Create a folder for the tile Editor hooks
